# [update] stopped using pointers for params