# Overview
![スクリーンショット 2022-11-12 12 57 44](https://user-images.githubusercontent.com/13291041/201456048-3cf605b8-ea73-4013-a564-39f81f72f278.png)

![スクリーンショット 2022-11-12 12 58 40](https://user-images.githubusercontent.com/13291041/201456083-b754597a-0a00-4a0c-a846-373a8d85fd1a.png)


# Changes
- Stopped using pointers for params
- Fix tests

Before
```sh
go test -bench=. -cpu=1 -benchmem -count=3
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1         9479496               130.1 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         9238120               129.5 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         9217893               130.2 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4274468               280.7 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4251072               281.0 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4269682               281.9 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2536338               472.8 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2542339               467.1 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2573580               467.8 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1774546               673.3 ns/op           392 B/op          7 allocs/op
BenchmarkWildCard1       1817253               668.5 ns/op           392 B/op          7 allocs/op
BenchmarkWildCard1       1776864               665.7 ns/op           392 B/op          7 allocs/op
BenchmarkWildCard5        471318              2510 ns/op             728 B/op         18 allocs/op
BenchmarkWildCard5        465484              2570 ns/op             728 B/op         18 allocs/op
BenchmarkWildCard5        475708              2506 ns/op             728 B/op         18 allocs/op
BenchmarkWildCard10       252789              4745 ns/op            1136 B/op         29 allocs/op
BenchmarkWildCard10       252019              4729 ns/op            1136 B/op         29 allocs/op
BenchmarkWildCard10       264942              4687 ns/op            1136 B/op         29 allocs/op
BenchmarkRegexp1         1856025               648.6 ns/op           392 B/op          7 allocs/op
BenchmarkRegexp1         1839674               647.2 ns/op           392 B/op          7 allocs/op
BenchmarkRegexp1         1809133               647.7 ns/op           392 B/op          7 allocs/op
BenchmarkRegexp5          497331              2396 ns/op             728 B/op         18 allocs/op
BenchmarkRegexp5          496726              2378 ns/op             728 B/op         18 allocs/op
BenchmarkRegexp5          524958              2407 ns/op             728 B/op         18 allocs/op
BenchmarkRegexp10         266846              4500 ns/op            1136 B/op         29 allocs/op
BenchmarkRegexp10         266263              4584 ns/op            1136 B/op         29 allocs/op
BenchmarkRegexp10         265555              4497 ns/op            1136 B/op         29 allocs/op
PASS
ok      github.com/bmf-san/goblin       40.125s
```

After
operation times and allocs/op has been improved, but ns/op and B/op are a little suspicious. 🤔 

```sh
go test -bench=. -cpu=1 -benchmem -count=3
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1         9416574               129.5 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         8315736               128.9 ns/op            32 B/op          1 allocs/op
BenchmarkStatic1         9438574               128.1 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4334941               278.0 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4338223               277.8 ns/op            96 B/op          1 allocs/op
BenchmarkStatic5         4298078               277.4 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2554701               466.5 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2543463               464.2 ns/op           176 B/op          1 allocs/op
BenchmarkStatic10        2565387               466.1 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1862164               645.0 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard1       1858161               647.9 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard1       1824388               662.0 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard5        476378              2725 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard5        492577              2580 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard5        396484              2753 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard10       257245              4917 ns/op            1560 B/op         19 allocs/op
BenchmarkWildCard10       227640              4790 ns/op            1560 B/op         19 allocs/op
BenchmarkWildCard10       256834              5053 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp1         1898145               635.4 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp1         1812205               640.4 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp1         1909863               670.1 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp5          424128              2494 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp5          519808              2550 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp5          507451              2448 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp10         262260              4419 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp10         259952              4362 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp10         283237              4340 ns/op            1560 B/op         19 allocs/op
PASS
ok      github.com/bmf-san/goblin       39.987s
```

# Impact range
No specification changes.
# Operational Requirements
N/A

# Related Issue
N/A

# Supplement
N/A